display alert deletion validation

+++ Views/ConnectionsPage.xaml.cs
@@ -1,5 +1,6 @@
 using CrossPath.ViewModels;
 using Microsoft.Maui.Controls;
+using System.Diagnostics;
 
 namespace CrossPath.Views;
 
@@ -14,9 +15,18 @@ public partial class ConnectionsPage : ContentPage
 
     private void DelBtn_Clicked(object sender, EventArgs e)
     {
-        File.Delete(Path.Combine(FileSystem.AppDataDirectory, "Connections.txt"));
-        File.Delete(Path.Combine(FileSystem.AppDataDirectory, "MapPins.txt"));
-        Data.ConnectionsCollection.Clear();
-        Data.MapPins.Clear();
+        AlertAsync();
+    }
+
+    private async Task AlertAsync() 
+    {
+        bool answer = await DisplayAlert("Delete?", "Would you like to delete all your connections?", "Yes", "No");
+        if (answer)
+        {
+            File.Delete(Path.Combine(FileSystem.AppDataDirectory, "Connections.txt"));
+            File.Delete(Path.Combine(FileSystem.AppDataDirectory, "MapPins.txt"));
+            Data.ConnectionsCollection.Clear();
+            Data.MapPins.Clear();
+        }
     }
 }
